+++ src/roiextractors/extractors/simaextractor/simasegmentationextractor.py
@@ -200,7 +200,7 @@ class SimaSegmentationExtractor(SegmentationExtractor):
         tuple
             The frame shape as (height, width).
         """
-        return self._image_masks.shape[0:2]
+        return self._roi_masks.field_of_view_shape
 
     def get_image_size(self):
         warnings.warn(